touch  ups
