+++ homeassistant/components/logbook/helpers.py
@@ -5,9 +5,16 @@ from collections.abc import Callable
 from typing import Any
 
 from homeassistant.components.sensor import ATTR_STATE_CLASS
-from homeassistant.const import ATTR_UNIT_OF_MEASUREMENT, EVENT_LOGBOOK_ENTRY
-from homeassistant.core import HomeAssistant, State
+from homeassistant.const import (
+    ATTR_DEVICE_ID,
+    ATTR_ENTITY_ID,
+    ATTR_UNIT_OF_MEASUREMENT,
+    EVENT_LOGBOOK_ENTRY,
+    EVENT_STATE_CHANGED,
+)
+from homeassistant.core import CALLBACK_TYPE, Event, HomeAssistant, State, callback
 from homeassistant.helpers import device_registry as dr, entity_registry as er
+from homeassistant.helpers.event import async_track_state_change_event
 
 from .const import (
     ALL_EVENT_TYPES_EXCEPT_STATE_CHANGED,
@@ -53,6 +60,69 @@ def async_filter_entities(hass: HomeAssistant, entity_ids: list[str]) -> list[st
         for entity_id in entity_ids
         if not is_entity_id_filtered(hass, ent_reg, entity_id)
     ]
+
+
+@callback
+def async_subscribe_events(
+    hass: HomeAssistant,
+    subscriptions: list[CALLBACK_TYPE],
+    target: Callable[[Event], None],
+    event_types: tuple[str, ...],
+    entity_ids: list[str] | None,
+    device_ids: list[str] | None,
+) -> None:
+    """Subscribe to events for the entities and devices or all."""
+    ent_reg = er.async_get(hass)
+    if entity_ids or device_ids:
+        entity_ids_set = set(entity_ids) if entity_ids else set()
+        device_ids_set = set(device_ids) if device_ids else set()
+
+        @callback
+        def _forward_events_filtered(event: Event) -> None:
+            event_data = event.data
+            if (
+                entity_ids_set and event_data.get(ATTR_ENTITY_ID) in entity_ids_set
+            ) or (device_ids_set and event_data.get(ATTR_DEVICE_ID) in device_ids_set):
+                target(event)
+
+        event_forwarder = _forward_events_filtered
+    else:
+
+        @callback
+        def _forward_events(event: Event) -> None:
+            target(event)
+
+        event_forwarder = _forward_events
+
+    for event_type in event_types:
+        subscriptions.append(
+            hass.bus.async_listen(event_type, event_forwarder, run_immediately=True)
+        )
+
+    @callback
+    def _forward_state_events_filtered(event: Event) -> None:
+        if event.data.get("old_state") is None or event.data.get("new_state") is None:
+            return
+        state: State = event.data["new_state"]
+        if not is_state_filtered(ent_reg, state):
+            target(event)
+
+    if entity_ids:
+        subscriptions.append(
+            async_track_state_change_event(
+                hass, entity_ids, _forward_state_events_filtered
+            )
+        )
+        return
+
+    # We want the firehose
+    subscriptions.append(
+        hass.bus.async_listen(
+            EVENT_STATE_CHANGED,
+            _forward_state_events_filtered,
+            run_immediately=True,
+        )
+    )
 
 
 def async_determine_event_types(
